split concrete into a separate stack
